[DONE] Admin extension download API TC

+++ Server/tests/views/apply/extension.py
@@ -2,7 +2,7 @@ from datetime import datetime
 import json
 import unittest2 as unittest
 
-from tests.views import account_student
+from tests.views import account_admin, account_student
 
 from server import app
 from utils.extension_meta import *
@@ -12,10 +12,13 @@ class TestExtension(unittest.TestCase):
     def setUp(self):
         self.client = app.test_client()
 
+        account_admin.create_fake_account()
         account_student.create_fake_account()
-        self.access_token = account_student.get_access_token(self.client)
+        self.admin_access_token = account_admin.get_access_token(self.client)
+        self.student_access_token = account_student.get_access_token(self.client)
 
     def tearDown(self):
+        account_admin.remove_fake_account()
         account_student.remove_fake_account()
 
     def testA_applyExtension11(self):
@@ -23,24 +26,19 @@ class TestExtension(unittest.TestCase):
         TC about extension apply 11
 
         1. Check 'unapplied'
-        2. Check 'unauthorized on extension apply'
-        3. Check 'apply succeed'
-        4. Check 'apply status'
+        2. Check 'apply succeed'
+        3. Check 'apply status'
         """
-        rv = self.client.get('/extension/11', headers={'Authorization': self.access_token})
+        rv = self.client.get('/extension/11', headers={'Authorization': self.student_access_token})
         self.assertEqual(rv.status_code, 204)
         # Unapplied check
 
-        rv = self.client.post('/extension/11')
-        self.assertEqual(rv.status_code, 401)
-        # Unauthorized check
-
-        rv = self.client.post('/extension/11', headers={'Authorization': self.access_token}, data={'class': 1, 'seat': 21})
+        rv = self.client.post('/extension/11', headers={'Authorization': self.student_access_token}, data={'class': 1, 'seat': 21})
         if APPLY_START < datetime.now().time() < APPLY_END_11:
             self.assertEqual(rv.status_code, 201)
             # Apply success
 
-            rv = self.client.get('/extension/11', headers={'Authorization': self.access_token})
+            rv = self.client.get('/extension/11', headers={'Authorization': self.student_access_token})
             self.assertEqual(rv.status_code, 200)
 
             data = json.loads(rv.data.decode())
@@ -53,19 +51,14 @@ class TestExtension(unittest.TestCase):
         """
         TC about withdraw extension apply 11
 
-        1. Check 'unauthorized on withdraw'
-        2. Check 'withdraw succeed'
-        3. Check 'unapplied'
+        1. Check 'withdraw succeed'
+        2. Check 'unapplied'
         """
-        rv = self.client.delete('/extension/11')
-        self.assertEqual(rv.status_code, 401)
-        # Unauthorized check
-
-        rv = self.client.delete('/extension/11', headers={'Authorization': self.access_token})
+        rv = self.client.delete('/extension/11', headers={'Authorization': self.student_access_token})
         self.assertEqual(rv.status_code, 200)
         # Withdraw success
 
-        rv = self.client.get('/extension/11', headers={'Authorization': self.access_token})
+        rv = self.client.get('/extension/11', headers={'Authorization': self.student_access_token})
         self.assertEqual(rv.status_code, 204)
         # Unapplied check
 
@@ -74,24 +67,19 @@ class TestExtension(unittest.TestCase):
         TC about extension apply 12
 
         1. Check 'unapplied'
-        2. Check 'unauthorized on extension apply'
-        3. Check 'apply succeed'
-        4. Check 'apply status'
+        2. Check 'apply succeed'
+        3. Check 'apply status'
         """
-        rv = self.client.get('/extension/12', headers={'Authorization': self.access_token})
+        rv = self.client.get('/extension/12', headers={'Authorization': self.student_access_token})
         self.assertEqual(rv.status_code, 204)
         # Unapplied check
 
-        rv = self.client.post('/extension/12')
-        self.assertEqual(rv.status_code, 401)
-        # Unauthorized check
-
-        rv = self.client.post('/extension/12', headers={'Authorization': self.access_token}, data={'class': 1, 'seat': 21})
+        rv = self.client.post('/extension/12', headers={'Authorization': self.student_access_token}, data={'class': 1, 'seat': 21})
         if APPLY_START < datetime.now().time() < APPLY_END_12:
             self.assertEqual(rv.status_code, 201)
             # Apply success
 
-            rv = self.client.get('/extension/12', headers={'Authorization': self.access_token})
+            rv = self.client.get('/extension/12', headers={'Authorization': self.student_access_token})
             self.assertEqual(rv.status_code, 200)
 
             data = json.loads(rv.data.decode())
@@ -104,24 +92,27 @@ class TestExtension(unittest.TestCase):
         """
         TC about withdraw extension apply 12
 
-        1. Check 'unauthorized on withdraw'
-        2. Check 'withdraw succeed'
-        3. Check 'unapplied'
+        1. Check 'withdraw succeed'
+        2. Check 'unapplied'
         """
-        rv = self.client.delete('/extension/12')
-        self.assertEqual(rv.status_code, 401)
-        # Unauthorized check
-
-        rv = self.client.delete('/extension/12', headers={'Authorization': self.access_token})
+        rv = self.client.delete('/extension/12', headers={'Authorization': self.student_access_token})
         self.assertEqual(rv.status_code, 200)
         # Withdraw success
 
-        rv = self.client.get('/extension/12', headers={'Authorization': self.access_token})
+        rv = self.client.get('/extension/12', headers={'Authorization': self.student_access_token})
         self.assertEqual(rv.status_code, 204)
         # Unapplied check
 
     def testE_downloadExtension11(self):
-        pass
+        """
+        TC about download extension apply 11
+        """
+        rv = self.client.get('/admin/extension/11', headers={'Authorization': self.admin_access_token})
+        self.assertEqual(rv.status_code, 200)
 
     def testF_downloadExtension12(self):
-        pass
+        """
+        TC about download extension apply 12
+        """
+        rv = self.client.get('/admin/extension/12', headers={'Authorization': self.admin_access_token})
+        self.assertEqual(rv.status_code, 200)
